feat: add experimental RelativeTime

Copilot-Session: bfa22b31-cbe2-4b07-8add-f302af4c1a2b
