+++ .changeset/react-first-relative-time.md
@@ -1,5 +1,5 @@
 ---
-'@primer/react': major
+'@primer/react': minor
 ---
 
-RelativeTime: Render a native time element without the custom-element dependency. The custom-element lifecycle, `update`, and `as` props are no longer supported.
+Add an experimental React-only RelativeTime implementation, available through `@primer/react/experimental`. The established RelativeTime component remains backed by the custom element by default; the new implementation can be enabled internally with `primer_react_relative_time`.
