+++ packages/react/package.json
@@ -76,6 +76,7 @@
   },
   "dependencies": {
     "@github/mini-throttle": "^2.1.1",
+    "@github/relative-time-element": "^5.0.0",
     "@lit-labs/react": "1.2.1",
     "@oddbird/popover-polyfill": "^0.5.2",
     "@primer/behaviors": "^1.10.3",
