+++ packages/react/src/FeatureFlags/DefaultFeatureFlags.ts
@@ -9,4 +9,5 @@ export const DefaultFeatureFlags = FeatureFlagScope.create({
   primer_react_action_list_item_gap: false,
   primer_react_timeline_list_semantics: false,
   primer_react_merged_forwarded_refs: false,
+  primer_react_relative_time: false,
 })
