+++ packages/react/src/RelativeTime/index.ts
@@ -1,2 +1,4 @@
 export {default} from './RelativeTime'
 export type {RelativeTimeProps} from './RelativeTime'
+export {default as ExperimentalRelativeTime} from './ExperimentalRelativeTime'
+export type {ExperimentalRelativeTimeProps, ExperimentalRelativeTimeUpdatedEvent} from './ExperimentalRelativeTime'
